+++ database/migrations/2019_10_30_152539_remove_field_employee_id_to_users_table.php
@@ -29,7 +29,7 @@ class RemoveFieldEmployeeIdToUsersTable extends Migration
     {
         Schema::table('users', function (Blueprint $table) {
             if (!Schema::hasColumn('users', 'employee_id')) {
-                $table->integer('employee_id')->unsigned()->nullable()
+                $table->bigInteger('employee_id')->unsigned()->nullable()
                       ->comment('Identificador del empleado asociado al usuario');
             }
         });
