+++ modules/Accounting/Database/Migrations/2019_03_22_143604_create_accounting_account_converters_table.php
@@ -15,12 +15,12 @@ class CreateAccountingAccountConvertersTable extends Migration
     {
         if (!Schema::hasTable('accounting_account_converters')) {
             Schema::create('accounting_account_converters', function (Blueprint $table) {
-                $table->increments('id');
+                $table->bigIncrements('id');
 
-                $table->integer('accounting_account_id')->unsigned()->comment('llave foranea a registro en la tabla accounting_accounts');
+                $table->bigInteger('accounting_account_id')->unsigned()->comment('llave foranea a registro en la tabla accounting_accounts');
                 $table->foreign('accounting_account_id')->references('id')->on('accounting_accounts')->onDelete('cascade')->comment('llave foranea a registro en la tabla accounting_accounts');
 
-                $table->integer('budget_account_id')->unsigned()->comment('llave foranea a registro en la tabla budget_accounts');
+                $table->bigInteger('budget_account_id')->unsigned()->comment('llave foranea a registro en la tabla budget_accounts');
                 $table->foreign('budget_account_id')->references('id')->on('budget_accounts')->onDelete('cascade')->comment('llave foranea a registro en la tabla budget_accounts');
 
                 $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
