+++ modules/Accounting/Database/Migrations/2019_10_02_125937_create_accounting_entries_table.php
@@ -24,18 +24,18 @@ class CreateAccountingEntriesTable extends Migration
             $table->float('tot_debit', 30, 10)->comment('Monto asignado al Debe total del asiento');
             $table->float('tot_assets', 30, 10)->comment('Monto asignado al Haber total del Asiento');
 
-            $table->integer('accounting_entry_categories_id')->unsigned()->nullable()->comment('id de la categoria u origen por el cual se genero el asiento contable');
+            $table->bigInteger('accounting_entry_categories_id')->unsigned()->nullable()->comment('id de la categoria u origen por el cual se genero el asiento contable');
             $table->foreign('accounting_entry_categories_id')->references('id')->on('accounting_entry_categories')->onDelete('cascade')->comment('id de la categoria u origen por el cual se genero el asiento contable');
 
-            $table->integer('currency_id')->unsigned()->nullable()
+            $table->bigInteger('currency_id')->unsigned()->nullable()
                 ->comment('id del tipo de moneda en que se guardar el asiento contable');
-                
+
             $table->foreign('currency_id')->references('id')->on('currencies')
                 ->onDelete('cascade')->comment('id del tipo de moneda en que se guardar el asiento contable');
 
             $table->boolean('approved')->default(false)->comment('Indica si el asiento contable fue verificado');
 
-            $table->integer('institution_id')->unsigned()->nullable()->comment('id de la institución que genero el asiento contable');
+            $table->bigInteger('institution_id')->unsigned()->nullable()->comment('id de la institución que genero el asiento contable');
             $table->foreign('institution_id')->references('id')->on('institutions')->onDelete('cascade')->comment('id de la institución que genero el asiento contable');
             $table->timestamps();
             $table->softDeletes()->comment('Fecha y hora en la que el registro fue eliminado');
